# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZenMon API Documentation",
        "description": "Monitoring Application - Alternative to Zabbix/Nagios",
        "contact": {
            "email": "admin@zenmon.local"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "ZenMon API Server"
        }
    ],
    "paths": {
...
        "/api/user-sessions": {
            "get": {
                "tags": [
                    "User Sessions"
                ],
                "summary": "Get list of user sessions",
                "description": "Returns paginated list of user sessions with filtering options",
                "operationId": "getUserSessionsList",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "hours",
                        "in": "query",
                        "description": "Get sessions from last N hours",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 720,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "expired_only",
                        "in": "query",
                        "description": "Show only expired sessions",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "ip_address",
                        "in": "query",
                        "description": "Filter by IP address",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user sessions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserSession"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Sessions"
                ],
                "summary": "Create new user session",
                "description": "Creates a new user session (login)",
                "operationId": "createUserSession",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "ip_address"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ip_address": {
                                        "type": "string",
                                        "example": "192.168.1.100"
                                    },
                                    "user_agent": {
                                        "type": "string",
                                        "example": "Mozilla/5.0..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Session created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserSession"
                                        },
                                        "session_token": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Session created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/user-sessions/active": {
            "get": {
                "tags": [
                    "User Sessions"
                ],
                "summary": "Get active user sessions",
                "description": "Returns list of currently active user sessions",
                "operationId": "getActiveUserSessions",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit number of results",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 200,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Active sessions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserSession"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total_active": {
                                                    "type": "integer"
                                                },
                                                "unique_users": {
                                                    "type": "integer"
                                                },
                                                "unique_ips": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserSession": {
                "title": "UserSession",
                "description": "User login session",
                "properties": {
                    "session_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "session_token": {
                        "type": "string",
                        "example": "abc123def456"
                    },
                    "login_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "last_activity_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "ip_address": {
                        "type": "string",
                        "example": "192.168.1.50"
                    },
                    "user_agent": {
                        "type": "string",
                        "example": "Mozilla/5.0..."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "login": {
                                "type": "string"
                            },
                            "full_name": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}